Add files via upload
The aim of this game is to reproduce a sequence of colours given by the programme. The program will display longer and longer chains of colours. The possible colours are blue, green, red and yellow.

The game begins when a sequence of three colours is displayed. While the sequence is displayed, the player cannot make any entries. If the sequence is complete, the player must repeat it using four buttons in the respective colours. If the answer is correct, another sequence is displayed which is one colour longer.

Progress is indicated by a level indicator. If the sequence is incorrect, the level drops by one, as indicated by a message in the game window.